Performance optimization for verification
Dependency verification may produce a number of
verification events which are only relevant if,
eventually, a fatal verification failure occurs.

If it's not the case, there will not be any
verification failure, therefore it's unnecessary
to perform formatting of an error message.

Previously, we would always check for verification
failures even if, in the end, there would only
be non fatal ones, which slows down IDE syncing.